+++ site/pages/PostPage.tsx
@@ -1,21 +1,21 @@
 import path from "node:path";
-import type { RootPageFn, Store } from "qrilljs/core";
+import type { RootNodeFn, Store } from "qrilljs/core";
 import { element, registerRootPage, S_2XLARGE, S_MEDIUM, style, W_MEDIUM } from "qrilljs/core";
 import { getMarkdown, listFiles, markdownToHtml } from "qrilljs/server";
-import { article } from "../../components/module/article.tsx";
-import { toc } from "../../components/module/toc.tsx";
-import { page } from "../../components/pages/page.tsx";
-import { navitem, postFmSchema, posts_dir, site } from "../../site.config.ts";
+import { article } from "../components/module/article.tsx";
+import { toc } from "../components/module/toc.tsx";
+import { page } from "../components/pages/page.tsx";
+import { navitem, postFmSchema, posts_dir, site } from "../site.config.ts";
 
-type RootParameter = {
+type PostPageParameter = {
     slug: string;
 };
 
-export async function rootPageFnParameters(): Promise<RootParameter[]> {
+export async function rootPageFnParameters(): Promise<PostPageParameter[]> {
     return listFiles(posts_dir, ".md").map((y) => ({ slug: path.basename(y, ".md") }));
 }
 
-export default function Root(store: Store): RootPageFn<RootParameter> {
+export function PostPage(store: Store): RootNodeFn<PostPageParameter> {
     const Page = page(store);
     const PageMainArea = element(store, { tag: "main", name: "page-main-area" });
     const Article = article(store);