+++ site/pages/PostsPage.tsx
@@ -1,14 +1,12 @@
-import type { RootPageFn, Store } from "qrilljs/core";
+import type { RootNodeFn, Store } from "qrilljs/core";
 import { element, registerRootPage, S_MEDIUM, style, W_MEDIUM } from "qrilljs/core";
 import { getAllMarkdowns } from "qrilljs/server";
 import { page } from "../components/pages/page.tsx";
-import { hero } from "../components/sections/hero.tsx";
 import { summaries } from "../components/sections/summaries.tsx";
 import { navitem, postFmSchema, posts_dir, site } from "../site.config.ts";
 
-export default function Home(store: Store): RootPageFn<void> {
+export function PostsPage(store: Store): RootNodeFn<void> {
     const Page = page(store);
-    const Hero = hero(store);
     const PageMainArea = element(store, { tag: "main", name: "page-main-area" });
     const Summaries = summaries(store);
 
@@ -25,11 +23,10 @@ export default function Home(store: Store): RootPageFn<void> {
 
     return async () => {
         const posts = await getAllMarkdowns(posts_dir, postFmSchema);
-        const posts_sorted = posts.sort((a, b) => new Date(b.data.date).getTime() - new Date(a.data.date).getTime());
+        const posts_sorted = posts.sort((a, b) => new Date(a.data.date).getTime() - new Date(b.data.date).getTime());
 
         return (
             <Page title={site.name} description={site.description} lang={site.lang} name={site.name} navitem={navitem}>
-                <Hero />
                 <PageMainArea>
                     <Summaries posts={posts_sorted} />
                 </PageMainArea>